refactor(QA): Generalise BackupAcceptance and RestoreAcceptance tests
