+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/backup/GcsBackupAcceptanceIT.java
@@ -7,22 +7,13 @@
  */
 package io.camunda.zeebe.it.backup;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 import com.google.cloud.storage.BucketInfo;
-import feign.FeignException;
 import io.camunda.zeebe.backup.gcs.GcsBackupConfig;
 import io.camunda.zeebe.backup.gcs.GcsBackupStore;
 import io.camunda.zeebe.broker.system.configuration.backup.BackupStoreCfg.BackupStoreType;
 import io.camunda.zeebe.broker.system.configuration.backup.GcsBackupStoreConfig;
 import io.camunda.zeebe.broker.system.configuration.backup.GcsBackupStoreConfig.GcsBackupStoreAuth;
 import io.camunda.zeebe.client.ZeebeClient;
-import io.camunda.zeebe.management.backups.BackupInfo;
-import io.camunda.zeebe.management.backups.PartitionBackupInfo;
-import io.camunda.zeebe.management.backups.StateCode;
-import io.camunda.zeebe.management.backups.TakeBackupResponse;
-import io.camunda.zeebe.qa.util.actuator.BackupActuator;
 import io.camunda.zeebe.qa.util.cluster.TestApplication;
 import io.camunda.zeebe.qa.util.cluster.TestCluster;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
@@ -31,14 +22,9 @@ import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.qa.util.testcontainers.GcsContainer;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
-import java.time.Duration;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.assertj.core.api.InstanceOfAssertFactories;
-import org.assertj.core.groups.Tuple;
-import org.awaitility.Awaitility;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
@@ -56,7 +42,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-final class GcsBackupAcceptanceIT {
+final class GcsBackupAcceptanceIT implements BackupAcceptance {
 
   private static final String BUCKET_NAME = RandomStringUtils.randomAlphabetic(10).toLowerCase();
 
@@ -97,79 +83,9 @@ final class GcsBackupAcceptanceIT {
     client = cluster.newClientBuilder().build();
   }
 
-  @Test
-  void shouldTakeBackup() {
-    // given
-    final var actuator = BackupActuator.ofAddress(cluster.availableGateway().monitoringAddress());
-    client.newPublishMessageCommand().messageName("name").correlationKey("key").send().join();
-
-    // when
-    final var response = actuator.take(1);
-
-    // then
-    assertThat(response).isInstanceOf(TakeBackupResponse.class);
-    waitUntilBackupIsCompleted(actuator, 1L);
-  }
-
-  private static void waitUntilBackupIsCompleted(
-      final BackupActuator actuator, final long backupId) {
-    Awaitility.await("until a backup exists with the id %d".formatted(backupId))
-        .atMost(Duration.ofSeconds(60))
-        .ignoreExceptions() // 404 NOT_FOUND throws exception
-        .untilAsserted(
-            () -> {
-              final var status = actuator.status(backupId);
-              assertThat(status)
-                  .extracting(BackupInfo::getBackupId, BackupInfo::getState)
-                  .containsExactly(backupId, StateCode.COMPLETED);
-              assertThat(status.getDetails())
-                  .flatExtracting(PartitionBackupInfo::getPartitionId)
-                  .containsExactlyInAnyOrder(1, 2);
-            });
-  }
-
-  @Test
-  void shouldListBackups() {
-    // given
-    final var actuator = BackupActuator.ofAddress(cluster.availableGateway().monitoringAddress());
-    client.newPublishMessageCommand().messageName("name").correlationKey("key").send().join();
-
-    // when
-    actuator.take(1);
-    actuator.take(2);
-
-    waitUntilBackupIsCompleted(actuator, 1L);
-    waitUntilBackupIsCompleted(actuator, 2L);
-
-    // then
-    final var status = actuator.list();
-    assertThat(status)
-        .hasSize(2)
-        .extracting(BackupInfo::getBackupId, BackupInfo::getState)
-        .containsExactly(
-            Tuple.tuple(1L, StateCode.COMPLETED), Tuple.tuple(2L, StateCode.COMPLETED));
-  }
-
-  @Test
-  void shouldDeleteBackup() {
-    // given
-    final var actuator = BackupActuator.ofAddress(cluster.availableGateway().monitoringAddress());
-    final long backupId = 1;
-    actuator.take(backupId);
-    waitUntilBackupIsCompleted(actuator, backupId);
-
-    // when
-    actuator.delete(backupId);
-
-    // then
-    Awaitility.await("Backup is deleted")
-        .timeout(Duration.ofSeconds(30))
-        .untilAsserted(
-            () ->
-                assertThatThrownBy(() -> actuator.status(backupId))
-                    .asInstanceOf(InstanceOfAssertFactories.type(FeignException.class))
-                    .extracting(FeignException::status)
-                    .isEqualTo(404));
+  @Override
+  public TestCluster getTestCluster() {
+    return cluster;
   }
 
   private void configureBroker(final TestStandaloneBroker broker) {
